Add files via upload
Given a time module and a cycle module, I created (mimicked) a sample of a traffics light cycle using time delay and a for loop. I also plan to use the time delay function in python to create code that that moves your cursor to random places across the screen.